+++ BidmadPluginSample/Assets/Bidmad/Scripts/AdSample/BannerAdSample.cs
@@ -14,7 +14,8 @@ public class BannerAdSample : MonoBehaviour
     public void LoadBannerAd()
     {
 #if UNITY_ANDROID
-        banner = new BidmadBanner("944fe870-fa3a-4d1b-9cc2-38e50b2aed43", 130);
+        banner = new BidmadBanner("944fe870-fa3a-4d1b-9cc2-38e50b2aed43", 130); // Auto Center alignment
+        // banner = new BidmadBanner("944fe870-fa3a-4d1b-9cc2-38e50b2aed43", 0, 130); // If you need to set position x.
 #elif UNITY_IOS
         banner = new BidmadBanner("1c3e3085-333f-45af-8427-2810c26a72fc", 130);
 #endif