+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Version 2.5.3
+#### Changes
+- Add interface for setting banner position X in Android
+- Fix Bugs
+#### Integration Bidmad SDK version 
+- Android : bidmad-androidx-g 1.0.7
+- iOS : BidmadSDK.framework 2.3.4
+
 # Version 2.5.2
 #### Changes
 - Admob SDK 8.0.0 Update